refactor: Create module with shared utilities
