+++ common/src/main/kotlin/flank.common/Archive.kt
@@ -1,9 +1,15 @@
-package flank.scripts.utils
+package flank.common
 
 import org.rauschig.jarchivelib.ArchiveFormat
 import org.rauschig.jarchivelib.ArchiverFactory
 import org.rauschig.jarchivelib.CompressionType
 import java.io.File
+import java.io.FileInputStream
+import java.io.InputStream
+import java.util.zip.ZipEntry
+import java.util.zip.ZipFile
+import java.util.zip.ZipOutputStream
+
 
 fun File.extract(
     destination: File,
